backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
exifread==3.0.0
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
docker==7.0.0
httpx==0.26.0
transformers==4.37.0
torch==2.1.2
sentence-transformers==2.3.1
qdrant-client==1.7.3
numpy==1.26.4
python-dotenv==1.0.1
